requirements.txt
aiogram
python-dotenv
requests
instaloader
tiktok-downloader